+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Character actors gain -1 init bonus while not wearing armor.
+
 ## [0.2.33] - 2022-11-21
 ### Fixed
 - Missing dialog for ranged attacks (oops!)